保存まで実装

+++ web-api/app/models/User.scala
@@ -27,8 +27,10 @@ case class User(
 
    override def extract(rs: WrappedResultSet, n: scalikejdbc.ResultName[User]): User =
      autoConstruct(rs, n)
+
    def create(user: User)(implicit session: DBSession): Long =
      createWithAttributes(toNamedValues(user): _*)
-   def update(user: User)(implicit session: DBSession): Int =
+
+   def update(user: User)(implicit session: DBSession): Long =
      updateById(user.id.get).withAttributes(toNamedValues(user): _*)
  }
